Develop

+++ apps/ui/src/renderer/src/utils/createWineAppPipeline/index.ts
@@ -1,7 +1,7 @@
-import { ProcessStatus, ExitCode, WineAppMode } from '@constants/enums';
+import { ProcessStatus, ExitCode, ScriptOperation } from '@constants/enums';
 import { FilePath } from '@interfaces/FilePath';
+import { PipelineScript } from '@interfaces/PipelineScript';
 import { SpawnProcessArgs } from '@interfaces/SpawnProcessArgs';
-import { WineAppConfig } from '@interfaces/WineAppConfig';
 import { WineAppJob } from '@interfaces/WineAppJob';
 import { WineAppJobWithScript } from '@interfaces/WineAppJobWithScript';
 import { WineAppPipeline } from '@interfaces/WineAppPipeline';
@@ -11,46 +11,53 @@ import { WineAppStep } from '@interfaces/WineAppStep';
 import { clone } from '@utils/clone';
 import { createEnv } from '@utils/createEnv';
 import { createWineApp } from '@utils/createWineApp';
-import { execCommand } from '@utils/execCommand';
+import { dirExists } from '@utils/dirExists';
+import { downloadFile } from '@utils/downloadFile';
 import { fileExists } from '@utils/fileExists';
 import { readDirectory } from '@utils/readDirectory';
 import { readFileAsString } from '@utils/readFileAsString';
+import { writeBinaryFile } from '@utils/writeBinaryFile';
 import { writeFile } from '@utils/writeFile';
+import { findOutputPids } from '@utils/findOutputPids';
 import { v4 as uuid } from 'uuid';
+import { parsePath } from '@utils/parsePath';
+import { getRelativeDriveCPath } from '@utils/getRelativeDriveCPath';
+import { createSteamCli } from '@utils/createSteamCli';
 
 export const createWineAppPipeline = async (options: {
-  appConfig: WineAppConfig;
+  appName: string;
   debug?: boolean;
   outputEveryMs?: number;
   promptMainExeCallback?: (args: {
     appExecutables: Array<FilePath>;
     driveCPath: string;
   }) => Promise<string>;
-  mode: WineAppMode;
+  clients?: {
+    steamCli?: ReturnType<typeof createSteamCli>;
+  };
 }) => {
   const id = uuid();
-  const store = { outputEnabled: true, killAllProcesses: false, currentProcess: { pid: 0 } };
+  const store = { outputEnabled: true, killAllProcesses: false, currentProcess: { pids: '0' } };
   const env = createEnv();
+  const wineApp = await createWineApp(options.appName);
+  const appConfig = wineApp.getAppConfig();
   const {
-    iconURL,
-    iconFile,
-    artworkFile,
-    name,
-    engineVersion,
-    engineURLs,
+    engineVersion = '',
+    engineURLs = [],
     dxvkEnabled,
     winetricks,
     setupExecutableURL,
     setupExecutablePath,
-    appFolderPath
-  } = options.appConfig;
-
-  const wineApp = await createWineApp(name, { mode: options.mode });
+    appFolderPath,
+    pipelineScripts = []
+  } = appConfig;
+  const { clients: { steamCli } = {} } = options;
   const appEnv = wineApp.getWineEnv();
   const PIPELINE_CONFIG_JSON_PATH = `${appEnv.WINE_APP_DATA_PATH}/pipeline.json`;
+  const WINETRICKS_VERSION = winetricks?.version || '20260125';
 
   let pipelineConfig: WineAppPipelineConfig = {
-    appConfig: options.appConfig,
+    appConfig,
     jobs: [],
     status: ProcessStatus.Pending
   };
@@ -69,7 +76,7 @@ export const createWineAppPipeline = async (options: {
         })
       };
     } else {
-      pipelineConfig.jobs.push(job);
+      pipelineConfig = { ...pipelineConfig, jobs: [...pipelineConfig.jobs, job] };
     }
   };
 
@@ -77,7 +84,7 @@ export const createWineAppPipeline = async (options: {
     const foundJob = pipelineConfig.jobs.find((item) => item.name == jobName);
     if (foundJob?.steps) {
       foundJob.steps = foundJob.steps.map((item) => {
-        if (item.name == step.name) return { ...item, ...step };
+        if (item.id == step.id) return { ...item, ...step };
         return item;
       });
     }
@@ -92,7 +99,9 @@ export const createWineAppPipeline = async (options: {
   };
 
   const writePipelineConfig = async () => {
-    await writeFile(PIPELINE_CONFIG_JSON_PATH, JSON.stringify(pipelineConfig));
+    if (await dirExists(appEnv.WINE_APP_DATA_PATH)) {
+      await writeFile(PIPELINE_CONFIG_JSON_PATH, JSON.stringify(pipelineConfig));
+    }
   };
 
   const initJobs = async (jobs: WineAppJobWithScript[]) => {
@@ -125,7 +134,7 @@ export const createWineAppPipeline = async (options: {
     const entries = (await readDirectory(ENGINES_PATH))
       .filter((item) => item !== '.DS_Store')
       .map((item) => item.replace(/.tar.7z$/, ''));
-    return entries.includes(`${engineVersion}.tar.7z`);
+    return entries.includes(engineVersion);
   };
 
   const ENGINE_EXISTS = await checkEngineExists();
@@ -150,8 +159,10 @@ export const createWineAppPipeline = async (options: {
 
     for (const verb of verbs) {
       steps.push({
+        id: uuid(),
         name: `Running winetrick ${verb}`,
-        script: (args: SpawnProcessArgs) => wineApp.winetrick(verb, args, winetricks?.options),
+        script: (args: SpawnProcessArgs) =>
+          wineApp.winetrick({ verb, version: WINETRICKS_VERSION }, args, winetricks?.options),
         status: ProcessStatus.Pending,
         output: ''
       });
@@ -162,14 +173,22 @@ export const createWineAppPipeline = async (options: {
 
   const resetJobStepsStatus = (
     steps: WineAppJobWithScript['steps'],
-    onUpdate: ((status: WineAppPipelineStatus) => void) | undefined
+    onUpdate: ((status: WineAppPipelineStatus) => void) | undefined,
+    fromStepIndex?: number
   ) => {
+    let stepIndex = 0;
     for (const step of steps) {
-      if (step.status == ProcessStatus.Success) {
+      if (fromStepIndex && stepIndex < fromStepIndex) {
+        stepIndex++;
+        continue;
+      }
+
+      if (step.status == ProcessStatus.Success && fromStepIndex == undefined) {
         continue;
       }
 
       step.status = ProcessStatus.Pending;
+      step.output = '';
     }
 
     onUpdate?.({
@@ -179,36 +198,152 @@ export const createWineAppPipeline = async (options: {
     });
   };
 
+  const updateCurrentProcess = (output: string) => {
+    const pids = findOutputPids(output);
+
+    if (pids && pids !== store.currentProcess.pids) {
+      store.currentProcess.pids = pids;
+    }
+  };
+
+  const runPipelineScript = async (args: PipelineScript, spawnProcessArgs: SpawnProcessArgs) => {
+    const { operation } = args;
+    const WINE_DOWNLOADS_PATH = `${env.get().WINE_DOWNLOADS_PATH}`;
+
+    switch (operation) {
+      case ScriptOperation.DOWNLOAD: {
+        spawnProcessArgs.onStdOut?.('-----');
+        spawnProcessArgs.onStdOut?.('Download Started:');
+        const fileName = args.downloadName || args.url.split('/').pop() || '';
+        const target = `${WINE_DOWNLOADS_PATH}/${decodeURIComponent(fileName)}`;
+
+        if (await fileExists(target)) {
+          spawnProcessArgs.onStdOut?.('File already exists, skipping download.');
+        } else {
+          let percent: number | undefined = undefined;
+          const file = await downloadFile(args.url, (args) => {
+            if (percent !== args.percent) {
+              percent = args.percent;
+              spawnProcessArgs.onStdOut?.(`${percent}%`);
+            }
+          });
+          await writeBinaryFile(target, file);
+          spawnProcessArgs.onStdOut?.('Download Finished.');
+        }
+
+        spawnProcessArgs.onExit?.(0);
+        break;
+      }
+      case ScriptOperation.DECOMPRESS: {
+        const path = parsePath(args.path);
+        const from = `${WINE_DOWNLOADS_PATH}/${path}`;
+        const target = from.replace(/\.[^.]+$/, '');
+        return wineApp.spawnScript('extract', `"${from}" "${target}"`, spawnProcessArgs);
+      }
+      case ScriptOperation.COPY: {
+        const from = `${WINE_DOWNLOADS_PATH}/${parsePath(args.from)}`;
+        const target = `${appEnv.WINE_APP_DRIVE_C_PATH}/${parsePath(args.target)}`;
+        return wineApp.spawnScript('copy', `"${from}" "${target}"`, spawnProcessArgs);
+      }
+      case ScriptOperation.REMOVE: {
+        const path = `${appEnv.WINE_APP_DRIVE_C_PATH}/${parsePath(args.removePath)}`;
+        return wineApp.spawnScript('remove', `"${path}"`, spawnProcessArgs);
+      }
+      case ScriptOperation.RUN_WINDOWS_EXE: {
+        const exePath = `/${parsePath(args.baseExePath)}/${parsePath(args.exePath)}`
+          .replace('$HOME', env.get().HOME)
+          .replace('$WINE_APP_PREFIX_PATH', appEnv.WINE_APP_PREFIX_PATH);
+        return wineApp.runExe(`${exePath}`, spawnProcessArgs);
+      }
+      case ScriptOperation.SET_MAIN_EXE: {
+        const mainExePath = getRelativeDriveCPath(
+          `${appEnv.WINE_APP_DRIVE_C_PATH}/${parsePath(args.mainExePath)}`
+        );
+        await wineApp.saveMainExecutablePath({ path: mainExePath, flags: args.exeFlags });
+        spawnProcessArgs.onExit?.(0);
+        break;
+      }
+      case ScriptOperation.MOUNT_DISK_IMAGE: {
+        const diskImagePath = `${WINE_DOWNLOADS_PATH}/${parsePath(args.diskImagePath)}`;
+        return wineApp.spawnScript('mountDiskImage', `"${diskImagePath}"`, spawnProcessArgs);
+      }
+      case ScriptOperation.DOWNLOAD_STEAM_APP: {
+        const gameInstallDir = `${appEnv.WINE_APP_DRIVE_C_PATH}/Program Files (x86)/Steam/steamapps/common/${args.installDirName}`;
+        return steamCli?.downloadSteamApp(
+          { appId: args.steamAppId, gameInstallDir },
+          spawnProcessArgs
+        );
+      }
+      default:
+        return;
+    }
+
+    return;
+  };
+
+  const buildPipelineStepsFromScripts = () => {
+    let steps: Array<
+      WineAppStep & {
+        script: (args: SpawnProcessArgs) => Promise<{
+          pid: number;
+        } | void>;
+      }
+    > = [];
+
+    for (const script of pipelineScripts) {
+      steps = [
+        ...steps,
+        {
+          id: uuid(),
+          name: script.name,
+          output: '',
+          status: ProcessStatus.Pending,
+          script: async (args) => runPipelineScript(script, args)
+        }
+      ];
+    }
+    return steps;
+  };
+
   const concatDataToOutput = (data: string | number | null, output = '') =>
     `${output || ''}\n${data}`;
 
   const pipeline: WineAppPipeline = {
     _: {
       async std(jobName, action, step, data, updateProcess) {
-        options.debug && console.log(action, step.name);
-
         if (store.killAllProcesses) {
           updateProcess?.('exit');
           step.status = ProcessStatus.Cancelled;
           savePipelineConfigJobStep(jobName, step);
+          this.onUpdate?.({
+            pipelineId: id,
+            jobs: pipeline.jobs,
+            status: ProcessStatus.Cancelled
+          });
           return;
         }
 
         step.status = ProcessStatus.InProgress;
         step.output = concatDataToOutput(data, step.output);
+        updateCurrentProcess(String(data));
 
         if (data === ExitCode.SuccessfulExecution) {
           step.status = ProcessStatus.Success;
+          this.onUpdate?.({
+            pipelineId: id,
+            jobs: pipeline.jobs,
+            status: ProcessStatus.InProgress
+          });
         }
 
-        if (data === ExitCode.Error) {
+        if (data === ExitCode.Error || data === ExitCode.PermissionsError) {
           step.status = ProcessStatus.Error;
         }
 
         savePipelineConfigJobStep(jobName, step);
 
         handleOutput(() => {
-          options.debug && console.log(action, data);
+          options.debug && console.log(action, data, step.name);
           this.onUpdate?.({
             pipelineId: id,
             jobs: pipeline.jobs,
@@ -217,6 +352,7 @@ export const createWineAppPipeline = async (options: {
         });
       }
     },
+    readPipelineConfig,
     onUpdate(fn) {
       this._.onUpdate = (pipelineStatus) => fn(clone(pipelineStatus));
     },
@@ -228,32 +364,21 @@ export const createWineAppPipeline = async (options: {
         status: ProcessStatus.Cancelled
       }),
     kill: async () => {
-      const pid = store.currentProcess.pid;
-      pid && execCommand(`kill -9 ${pid}`);
+      const pids = store.currentProcess.pids;
+      pids && (await wineApp.execScript('killPids', `${pids}`));
       store.killAllProcesses = true;
+      savePipelineStatus(ProcessStatus.Cancelled);
+      await writePipelineConfig();
     },
     jobs: [
       {
         name: 'Create wine app',
         steps: [
-          {
-            name: 'Creating wine app',
-            script: (args) =>
-              wineApp.scaffold(
-                {
-                  appIconURL: iconURL,
-                  appIconFile: iconFile,
-                  appArtWorkFile: artworkFile
-                },
-                args
-              ),
-            status: ProcessStatus.Pending,
-            output: ''
-          },
           ...(ENGINE_EXISTS
             ? []
             : [
                 {
+                  id: uuid(),
                   name: 'Downloading wine engine',
                   script: (args: SpawnProcessArgs) =>
                     wineApp.downloadWineEngine(engineURLs, engineVersion, args),
@@ -262,12 +387,14 @@ export const createWineAppPipeline = async (options: {
                 }
               ]),
           {
+            id: uuid(),
             name: 'Extracting wine engine',
             script: (args) => wineApp.extractEngine(engineVersion, args),
             status: ProcessStatus.Pending,
             output: ''
           },
           {
+            id: uuid(),
             name: 'Generating wine prefix',
             script: (args) => wineApp.wineboot('', args),
             status: ProcessStatus.Pending,
@@ -276,8 +403,10 @@ export const createWineAppPipeline = async (options: {
           ...(dxvkEnabled
             ? [
                 {
+                  id: uuid(),
                   name: 'Enabling DXVK',
-                  script: (args: SpawnProcessArgs) => wineApp.winetrick('dxvk1102', args),
+                  script: (args: SpawnProcessArgs) =>
+                    wineApp.winetrick({ verb: 'dxvk1102', version: WINETRICKS_VERSION }, args),
                   status: ProcessStatus.Pending,
                   output: ''
                 }
@@ -287,8 +416,10 @@ export const createWineAppPipeline = async (options: {
           ...(setupExecutableURL
             ? [
                 {
+                  id: uuid(),
                   name: 'Downloading setup executable',
-                  script: () => wineApp.setSetupExe(setupExecutableURL),
+                  script: (args?: SpawnProcessArgs) =>
+                    wineApp.setSetupExe(setupExecutableURL, args),
                   status: ProcessStatus.Pending,
                   output: ''
                 }
@@ -297,10 +428,11 @@ export const createWineAppPipeline = async (options: {
           ...(setupExecutablePath
             ? [
                 {
+                  id: uuid(),
                   name: 'Running setup executable',
                   script: (args?: SpawnProcessArgs) => {
-                    setupExecutablePath && wineApp.setSetupExe(setupExecutablePath);
-                    return wineApp.runExe(wineApp.getAppConfig().setupExecutablePath || '', args);
+                    const exePath = wineApp.getAppConfig().setupExecutablePath || '';
+                    return wineApp.runExe(exePath, args);
                   },
                   status: ProcessStatus.Pending,
                   output: ''
@@ -310,6 +442,7 @@ export const createWineAppPipeline = async (options: {
           ...(appFolderPath
             ? [
                 {
+                  id: uuid(),
                   name: 'Copying windows application',
                   script: (args?: SpawnProcessArgs) => {
                     return wineApp.copyWindowsApplication(appFolderPath || '', args);
@@ -319,12 +452,16 @@ export const createWineAppPipeline = async (options: {
                 }
               ]
             : []),
+          ...buildPipelineStepsFromScripts(),
           {
+            id: uuid(),
             name: 'Configuring app executable',
-            script: async (args) => {
-              let executables = options.appConfig.executables || [];
+            script: async (args?: SpawnProcessArgs) => {
+              const appConfig = wineApp.getAppConfig();
+              let executables = appConfig.executables || [];
+              const mainExecutablePath = executables.find((item) => item.main)?.path || '';
 
-              if (!options.appConfig.executables?.length) {
+              if (!mainExecutablePath) {
                 let exePath = '';
 
                 if (options.promptMainExeCallback) {
@@ -340,7 +477,8 @@ export const createWineAppPipeline = async (options: {
                 executables = [{ path: exePath, main: true }];
               }
 
-              return wineApp.bundleApp({ executables, configId: options.appConfig.id }, args);
+              await wineApp.setExecutables({ executables });
+              args?.onExit?.(0);
             },
             status: ProcessStatus.Pending,
             output: ''
@@ -348,12 +486,27 @@ export const createWineAppPipeline = async (options: {
         ]
       }
     ],
-    async run() {
+    async run(args?: { fromJobIndex?: number; fromStepIndex?: number }) {
+      const { fromJobIndex, fromStepIndex } = args || {};
+      savePipelineStatus(ProcessStatus.InProgress);
+      await writePipelineConfig();
+      let jobIndex = 0;
+
       for (const job of pipeline.jobs) {
+        let stepIndex = 0;
+        if (fromJobIndex && jobIndex < fromJobIndex) {
+          jobIndex++;
+          continue;
+        }
+
         savePipelineJob(job);
-        resetJobStepsStatus(job.steps, this._.onUpdate);
+        resetJobStepsStatus(job.steps, this._.onUpdate, fromStepIndex);
 
         for (const step of job.steps) {
+          if (fromStepIndex && stepIndex < fromStepIndex) {
+            stepIndex++;
+            continue;
+          }
           if (step.status == ProcessStatus.Success) {
             continue;
           }
@@ -370,7 +523,7 @@ export const createWineAppPipeline = async (options: {
             continue;
           }
 
-          store.currentProcess = await step.script({
+          await step.script({
             onStdOut: (data, updateProcess) =>
               this._.std(job.name, 'stdOut', step, data, updateProcess),
             onStdErr: (data, updateProcess) =>
